Merged PR 19911: docs: Add a glossary of terms
# 🔍 Description

Add a glossary with common terms and their meaning when it comes to Trident.
We should always be consistent in using the terms defined here.

----
#### AI description  (iteration 1)
#### PR Classification
Documentation

#### PR Summary
This pull request adds a glossary of terms to the project documentation to help clarify specific terminology used within the project.
- Added `Glossary.md` in `docs/Reference` with definitions for terms such as "Dualboot", "Install", "Multiboot", "Servicing", and "Servicing Type".

Related work items: #8529