Merged PR 20020: Set rust-version in Cargo.toml and fix warning from newer rustc
# 🔍 Description

The `rust-version` field in Cargo.toml signals the oldest version of the compiler that the package is intended to be compiled with. This can be helpful because it prevents newer versions of clippy from suggesting changes that wouldn't work with the compiler version we're targeting.

# 🤔 Rationale

This change makes it slightly smoother to develop locally with newer versions of Rust even though our pipelines are still set to 1.72.0. That is desirable because newer compiler releases are faster and produce  better warnings.

You can update the currently installed version of rust with:
```
rustup update
```

And if you ever want to compile using a specific versions of rust, you can manually install them (`rustup toolchain add 1.72.1`) and then use this syntax:
```
cargo +1.72.1 check
```

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup and configuration update.

#### PR Summary
This pull request sets the Rust version in `Cargo.toml` and fixes a warning from a newer Rust compiler.
- `Cargo.toml`: Added `rust-version = "1.72.0"`.
- `src/engine/hooks.rs`: Fixed a warning by changing `collect::<Result<_, _>>()?` to `collect::<Result<(), _>>()?`.

Related work items: #8672