Merged PR 20051: Bug: Image for 3.0
# 🔍 Description

Change the image for 3.0 image to build trident.

# 🤔 Rationale

We build the trident version compatible with azurelinux 3.0 inside a container of azurelinux 3.0. The source where we get the image isn't working anymore.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request updates the base image source in the Dockerfile to fix an issue with the image for version 3.0.
- `Dockerfile.full`: Changed the base image source from `azurelinuxpreview.azurecr.io` to `mcr.microsoft.com`.

Related work items: #8723