Merged PR 20115: bug: Force systemd-repart to generate unique PTUUID for each disk/partition table
# 🔍 Description

This PR fixes a bug where we were setting '--seed=random' in the call to systemd-repart. Under the hood, systemd-repart always sets seed to 0 if that's the case, due to an upstream bug in their logic. As a result, when the host had 2+ disks, all the PTUUIDs generated for the disks/partition tables were the same. This caused confusion in the host status and to Trident overall. This PR fixes this issue by generating a random UUID and then passing it as the seed value, so that each PTUUID is hashed uniquely.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to ensure unique PTUUID generation for each disk/partition table.

#### PR Summary
This pull request addresses the issue where `systemd-repart` assigns the same UUID to all disks, causing problems with disk detection and UUID cache.
- Modified `osutils/src/repart.rs` to generate a random UUID for the `seed` argument, ensuring unique PTUUIDs for each disk/partition table.
- Minor debug message correction in `src/engine/bootentries.rs`.

Related work items: #8561