Merged PR 20123: Engineering: New stage to build the docker image
# 🔍 Description

This PR moves the task that builds the docker image to its own stage.

# 🤔 Rationale

The docker image is used to test trident deployment from a container, but currently it wasn't built in the prerelease pipeline (because we download the rpms from the ci feed instead of building them). Moving the task to its own stage allows to build it in parallel to the test images and doesn't increase runtime of the pipeline,

# 📝 Checks

- Pipeline run example: https://dev.azure.com/mariner-org/ECF/_build/results?buildId=629558&view=results

# Related PRs:
- https://dev.azure.com/mariner-org/ECF/_git/trident/pullrequest/20058 (container testing)

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces a new pipeline stage for building Docker images for the Trident project.
- Added `/.pipelines/templates/stages/build_docker_image/trident-container.yml` to define the new Docker image build stage.
- Removed Docker image build steps from `/.pipelines/templates/stages/trident_rpms/release.yml`.
- Updated `/.pipelines/templates/e2e-template.yml` to include the new Docker image build stage.