Merged PR 20267: engineering: Add new tests to trident-testing pipeline
# 🔍 Description

The [trident-testing pipeline](https://dev.azure.com/mariner-org/ECF/_build?definitionId=3262&_a=summary) is used to just run the tests. This is useful when just modifying the testing infrastructure as you save the time of building trident/images. This PR add the option to run the new test stages (servicing testing and vm testing with trident from a container).

This PR also adds to the pipeline the ability to select a specific run of a pipeline even if the run was not successful (the artifacts needed would still need to have been built on that run).

Example:
![image.png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/20267/attachments/image.png)

# 🤔 Rationale

It saves time when making changes just to the testing infrastructure. It also is useful to test specific artifacts built in other pipeline runs.

# 📝 Checks

- Example of run with the latest successful run of a branch: https://dev.azure.com/mariner-org/ECF/_build/results?buildId=635610&view=results
- Example of run with an specific run ID selected (which wasn't successful, but built the needed artifacts): https://dev.azure.com/mariner-org/ECF/_build/results?buildId=635634&view=results

----
#### AI description  (iteration 1)
#### PR Classification
New feature: Add new tests to the trident-testing pipeline.

#### PR Summary
This pull request enhances the trident-testing pipeline by adding new test configurations and improving artifact download mechanisms.
- `.pipelines/templates/test-template.yml`: Added parameters for pipeline run ID and new test configurations for VM containers and servicing tests.
- `.pipelines/templates/stages/download_artifacts/get-artifacts.yml`: Introduced parameters for specific run versions and pipeline run IDs, and added jobs for downloading various trident artifacts.
- `.pipelines/testing-trident.yml`: Updated parameters to include new test configurations and pipeline run ID.
- `.pipelines/templates/stages/deployment_testing/servicing/vm-testing.yml`: Added parameters to conditionally depend on the `DownloadTestingElements` stage.

Related work items: #8975