Merged PR 20282: Fixing a typo in the Makefile for the bin/trident-rpms-azl2.tar.gz target
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

The issue is a typo in the trident Makefile. The "bin/trident-rpms-azl2.tar.gz" target points to the "Dockerfile", but that file doesn't exist. It supposedly should point to "Dockerfile**.azl2**" file. The bug currently makes the "make all" command fail from the command line.

# 🤔 Rationale

Why is this PR needed?

The PR fixes the bug that causes "make all" to fail.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

At the moment, I solely built the code to validate "make all" passed. Since this is a build issue, I wanted to raise the PR and can do additional testing while the code is being reviewed.

# 📌 Follow-ups

TODO:

- n/a

# 🗒️ Notes

Fixing a typo in the Makefile for the bin/trident-rpms-azl2.tar.gz target

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request fixes a typo in the Makefile for the `bin/trident-rpms-azl2.tar.gz` target, ensuring the correct dependency is specified.
- `Makefile`: Corrected the dependency from `Dockerfile` to `Dockerfile.azl2` for the `bin/trident-rpms-azl2.tar.gz` target.

Related work items: #8981