Merged PR 20386: hack: Have netlisten ignore the first failure to get better output on 'rerun'...
# 🔍 Description

The last E2E pipeline run for 'rerun' failed. Unfortunately, netlisten lacks the maxFailures argument that netlaunch has.

Longer term it would be nice to add the same functionality to netlisten (or move everything over to netlaunch...) but this is a faster workaround.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to improve output on 'rerun' for E2E tests.

#### PR Summary
This pull request addresses a failure in the 'rerun' E2E test by ignoring the first failure from `phonehome` in `netlisten`.
- `tools/cmd/netlisten/main.go`: Added logic to ignore the first failure from `phonehome` to support the 'rerun' E2E test.

Related work items: #9083