Merged PR 20672: Bug: Download latest successful or partially successful trident runtime partitions
# 🔍 Description

This PR modifies the make rules to get the pipeline run id of the latest successful or partially successful run of the trident-ci pipeline. Same process, but now we query the latest 10 runs of the trident-ci pipeline. Then, we filter to only leave the successful or partially successful runs, and we select the latest one.

# 🤔 Rationale

Before we were getting only the latest successful run, but currently we are allowing the Deployment VM container Testing stage to fail to solve [Bug 9348](https://dev.azure.com/mariner-org/ECF/_workitems/edit/9348). As a result, we aren't downloading the latest run of trident-ci pipeline.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses an issue where the Trident runtime OS does not reach out to netlaunch after a reboot by ensuring the download of the latest successful or partially successful runtime partitions.
- `Makefile`: Modified the `download-trident-container-installer-iso` and `download-runtime-partition-images` targets to fetch the latest successful or partially successful pipeline runs instead of only successful ones.

Related work items: #9348