Merged PR 20671: engineering: Separate internal and external HostStatus representation
# 🔍 Description

Adds EngineContext as a internal representation of the state of the host. Initially, it contains the same contents as the public HostStatus (except servicing_state is removed), but we can consider caching additional information in it going forward.

# 🤔 Rationale

This will make it easier to expose a stable API for the HS and to cache system information while Trident is running

----
#### AI description  (iteration 1)
#### PR Classification
Design

#### PR Summary
This pull request separates the internal `EngineContext` state representation from the external `HostStatus` representation to improve code clarity and maintainability.
- `src/engine/boot/mod.rs`: Replaced `HostStatus` with `EngineContext` in various test cases.
- `src/engine/storage/image/mod.rs`: Updated references from `HostStatus` to `EngineContext` in block device path handling.
- `src/engine/mod.rs`: Introduced `EngineContext` struct with relevant fields and methods.
- `src/engine/storage/verity.rs`: Modified verity device setup functions to use `EngineContext`.
- `src/engine/bootentries.rs` and `src/engine/boot/esp.rs`: Updated functions to use `EngineContext` instead of `HostStatus`.

Related work items: #9105