Merged PR 20714: test: Have re-run test fail during post-configure script rather than post-provision
# 🔍 Description

Tests that Trident can rerun following a failing post-configure script. This is a more challenging case for Trident to handle because the failure happens within the chroot and after more provisioning steps have been completed.

The implementation itself uses a post-provision script to schedule the re-run and a post-configure script to crash the install on the first run. We cannot easily do everything inside the post-configure script because the chroot doesn't have access to the host's systemd instance, at least not directly. (It may be possible to establish a connection, but I wasn't able to figure that out in the limited amount of searching I tried.)

Related work items: #8470