Merged PR 20647: engineering: Add logic for tracking sizes of partitions and raid arrays as fields to calculate throughput
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

It will be helpful to track the throughput of encryption and raid operations as the sizes of the partitions/arrays heavily dictate the time it takes for the operations to complete. This PR adds some logic to handle tracking the size as a span field which can be used in a KQL query to calculate the throughput, not just the time for the operations.

# 🤔 Rationale

Why is this PR needed?

Help measure throughput of features like encryption and raid, not just time.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

Add logic for tracking sizes of partitions and raid arrays as fields

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds logic to track the sizes of partitions and RAID arrays to calculate throughput.
- `src/logging/tracestream.rs`: Added `on_record` function to handle storing fields with the visitor pattern and added `record_u64` method to `TraceEntryVisitor`.
- `src/engine/storage/encryption.rs`: Added logic to convert partition size from bytes to MiB and track total partition size, and updated tracing instrumentation to include `total_partition_size_mb`.

Related work items: #9382