Merged PR 20774: engineering: Update GRUB config for azl3 with OSModifier
# 🔍 Description

What is this PR about?
Use EMU to update GRUB config azl 3.0

1st time EMU invocation: `osmodifier::update_grub` under the hood, EMU reads verity/overlay/selinux/root device from original grub.cfg, then update (if exists) or add (if not exists) these values to /etc/default/grub, and then run grub2-mkconfig.

2nd time EMU invocation: `osmodifier::run()` pass verity/overlay/selinux/root device from engine context to EMU API to apply the values to /etc/default/grub, and run grub2-mkconfig.

Also increased the timeout minutes to 13 as testing rerun exceeded 11 minutes sometimes.

verified locally and the deployments and fts in pipeline are green

# 🤔 Rationale

Why is this PR needed?

For Azure Linux 3.0, grub.cfg is not directly edited but generated from /etc/default/grub, OS Modifier is used to apply the changes to /etc/default/grub and run grub-mkconfig.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

Moving forward, it'd be good to aggregate all the changes to /etc/default/grub, and invoke EMU once to make the changes.

# 🗒️ Notes

----
#### AI description  (iteration 1)
#### PR Classification
Code enhancement for updating GRUB configuration in Azure Linux 3.0 using OSModifier.

#### PR Summary
This pull request updates the GRUB configuration process for Azure Linux 3.0 by integrating the OSModifier tool, which simplifies and centralizes the configuration updates.
- `src/engine/boot/grub.rs`: Replaces manual GRUB configuration updates with calls to the OSModifier tool, adds handling for SELinux, verity, and overlay configurations.
- `src/engine/storage/verity.rs`: Commented out old logic for updating Azure Linux 3.0 GRUB config.
- `src/engine/boot/emuboot.rs`: Adds new file defining structures for boot configuration (SELinux, Overlay, Verity).
- `trident_api/src/config/host/os/mod.rs`: Implements `Display` trait for `SelinuxMode`.
- `osutils/src/osmodifier.rs`: Adds function to update GRUB using OSModifier.

Related work items: #8476