Merged PR 20936: doc: update docs to indicate trident will use the first esp found
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?
update docs

# 🤔 Rationale

Why is this PR needed?

The current approach involves iterating through all filesystems and selecting the first one that matches the characteristics of an ESP partition. This decision is based on the inability to determine the user's intent when multiple ESPs are present in their configuration, except when redundancy is configured.

Updating docs to reflect on what trident currently does.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

make doc changes to indicate that only the first esp will be honoured

Related work items: #9622