Merged PR 20950: engineering: Add timeout to execute other tasks in case of failure
# 🔍 Description

This PR adds a timeout to the task that checks that Trident can perform an update on the Servicing Testing stage.

# 🤔 Rationale

We need the timeout so in case of a failure, we have enough time to publish the OS (last task in the template).

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to address a failing test in the CI pipeline.

#### PR Summary
This pull request adds a timeout to a specific task in the CI pipeline to prevent indefinite execution in case of failure.
- `/.pipelines/templates/stages/testing_servicing/vm-testing.yml`: Added a 5-minute timeout to the "Check that Trident can perform the update" task.

Related work items: #9737