Merged PR 20684: Remove local config
# 🔍 Description

This PR removes the concept of a local config and makes Trident directly deal with HC files where needed. The allowed operations are now passed as command line options.

In order to know whether Trident is running from the provisioning OS or the runtime OS, the current environment is tracked in the host status.

This PR makes `--config` a required argument to initiate an install or update operation, but uses /etc/trident/config.yaml as a fallback if not provided. Ideally, we'd remove the fallback and require users to affirmatively pass a HC if they want to do provisioning/update. However, that's somewhat disruptive because it means you cannot keep the same systemd unit file for both provisioning and post-reboot committing of the update.

Must merge at the same time as test-images PR: https://dev.azure.com/mariner-org/ECF/_git/test-images/pullrequest/21018

# 🤔 Rationale

Users have been confused about the difference between a local config and a host config. This also avoids having to decide on a stable API for the local config.

# Passing pipeline
https://dev.azure.com/mariner-org/ECF/_build/results?buildId=674382&view=results
![image.png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/20684/attachments/image.png)
----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
This pull request removes local configuration handling and refactors the code to use a more centralized configuration approach.
- `src/main.rs`: Removed `LocalConfigFile` and updated command structures to use `HostConfigurationSource`.
- `src/lib.rs`: Refactored `Trident` struct to eliminate local config handling and updated methods to use `HostConfigurationSource`.
- `src/validation.rs`: Removed validation functions related to local configuration files.
- `src/engine/management.rs`: Removed functions and code related to creating and managing local configuration files.

Related work items: #9429