Merged PR 21022: engineering: Write and persist metrics to local file on ROS
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

This PR enables writing tracing metrics to a separate metrics file that persists across the MOS and ROS. This will give us the ability to collect metrics from production runs of Trident for various servicing operations.

Leverages the logic that already exists for persisting background logs and adds metrics file to the function since required logic is the same.

The decision to keep the metrics and log files separate was agreed upon by the crew to avoid the possibility of customers not wanting to share their logs and the file could get messy/complicated, compared to being able to explain much easier the purpose and contents of the Trident metrics files.

# 🤔 Rationale

Why is this PR needed?

Enables metric collection from production runs of Trident, without this logic we won't have any production data for KPIs/dashboards.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

Write and persist metrics to local file

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds functionality to write and persist tracing metrics to a local file on the host system.
- `src/engine/mod.rs`: Added logic to copy metrics file to a new location and updated documentation comments.
- `src/logging/tracestream.rs`: Implemented methods to create and write metrics to a local file.
- `src/lib.rs`: Defined a constant for the metrics file path.

Related work items: #9821